requirements.txt
pillow==10.4.0
tk==0.1.0
numpy==2.1.1